chore: rename tests with mixed direct and pooled stacking
