+++ tests/integration/pox/stacking/direct-and-pooled-increase-as-pool-operator.spec.ts
@@ -16,11 +16,11 @@ import {
 import { broadcastStackSTX } from "../helpers-direct-stacking";
 import {
   broadcastDelegateSTX,
-  broadcastDelegateStackIncrease,
-  broadcastStackAggregationCommitIndexed,
+  broadcastDelegateStackSTX,
+  broadcastStackAggregationCommitIndexed
 } from "../helpers-pooled-stacking";
 
-describe("testing mixed direct and pooled stacking under epoch 2.1", () => {
+describe("testing direct stacker as pool operator without auto-unlock under epoch 2.1", () => {
   let orchestrator: DevnetNetworkOrchestrator;
   let timeline = {
     epoch_2_0: 100,
@@ -39,7 +39,7 @@ describe("testing mixed direct and pooled stacking under epoch 2.1", () => {
     orchestrator.terminate();
   });
 
-  it("using direct stacking and delegate-stack-increase with auto-unlock,", async () => {
+  it("using direct stacking and delegate-stack-increase should not result in auto-unlock", async () => {
     const network = new StacksTestnet({ url: orchestrator.getStacksNodeUrl() });
 
     // Wait for Stacks genesis block
@@ -50,12 +50,12 @@ describe("testing mixed direct and pooled stacking under epoch 2.1", () => {
     const blockHeight = timeline.pox_2_activation + 1;
     const cycles = 1;
 
-    // Faucet stacks 75m STX
+    // Alice stacks 80m STX
     let response = await broadcastStackSTX(
       2,
       network,
-      75_000_000_000_000,
-      Accounts.FAUCET,
+      80_000_000_000_000,
+      Accounts.WALLET_1,
       blockHeight,
       cycles,
       fee,
@@ -63,102 +63,109 @@ describe("testing mixed direct and pooled stacking under epoch 2.1", () => {
     );
     expect(response.error).toBeUndefined();
 
-    // Faucet delegates 1_100m to Bob
+    // Faucet delegates 999m to Alice address
     response = await broadcastDelegateSTX(
       2,
       network,
       Accounts.FAUCET,
       fee,
-      1,
-      1_100_000_000_000_000,
-      Accounts.WALLET_2
+      0,
+      999_000_000_000_000,
+      Accounts.WALLET_1
     );
     expect(response.error).toBeUndefined();
     let [block, tx] = await waitForStacksTransaction(
       orchestrator,
       response.txid
     );
-    expect(tx.success).toBeTruthy();
+    expect(tx.success, tx.result).toBeTruthy();
 
-    // Bob increase delegation by 920m
-    response = await broadcastDelegateStackIncrease(
+    // Alice locks 999m as pool operator
+    response = await broadcastDelegateStackSTX(
       2,
       network,
-      Accounts.WALLET_2,
+      Accounts.WALLET_1,
       fee,
-      0,
+      1,
       Accounts.FAUCET,
-      Accounts.WALLET_2,
-      920_000_000_000_000
+      999_000_000_000_000,
+      Accounts.WALLET_1,
+      blockHeight,
+      1
+    );
+    expect(response.error).toBeUndefined();
+    [block, tx] = await waitForStacksTransaction(orchestrator, response.txid);
+    expect(tx.success, tx.result).toBeTruthy();
+
+    // Alice commits 999m
+    response = await broadcastStackAggregationCommitIndexed(
+      2,
+      network,
+      Accounts.WALLET_1,
+      fee,
+      2,
+      Accounts.WALLET_1,
+      2
     );
     expect(response.error).toBeUndefined();
 
     [block, tx] = await waitForStacksTransaction(orchestrator, response.txid);
     expect(tx.success, tx.result).toBeTruthy();
 
-    // Bob commits 920m
-    response = await broadcastStackAggregationCommitIndexed(
-      2,
-      network,
-      Accounts.WALLET_2,
-      fee,
-      1,
-      Accounts.WALLET_2,
-      2
-    );
-    expect(response.error).toBeUndefined();
-
-    [block, tx] = await waitForStacksTransaction(orchestrator, response.txid);
-    expect(tx.result).toBe("(ok u1)");
-    expect(tx.success).toBeTruthy();
-
     let poxInfo = await getPoxInfo(network);
 
-    // Assert that the next cycle has 1079m STX locked
     expect(poxInfo.current_cycle.id).toBe(1);
-    expect(poxInfo.current_cycle.min_threshold_ustx).toBe(29_290_000_000_000);
-    expect(poxInfo.next_cycle.stacked_ustx).toBe(995_000_000_000_000);
+    // Assert that the next cycle has 990m STX locked
+    expect(poxInfo.next_cycle.stacked_ustx).toBe(1_079_000_000_000_000);
 
-    // Check Faucets's table entry
+    // Check Alice's table entry
     let poxAddrInfo0 = await readRewardCyclePoxAddressListAtIndex(
       network,
       2,
       0
     );
-    expect(poxAddrInfo0?.["total-ustx"]).toEqual(uintCV(75_000_000_000_000));
+    expect(poxAddrInfo0?.["total-ustx"]).toEqual(uintCV(80_000_000_000_000));
 
-    // Check Bob's table entry
+    // Check Alice's second table entry
     let poxAddrInfo1 = await readRewardCyclePoxAddressListAtIndex(
       network,
       2,
       1
     );
-    expect(poxAddrInfo1?.["total-ustx"]).toEqual(uintCV(920_000_000_000_000));
+    expect(poxAddrInfo1?.["total-ustx"]).toEqual(uintCV(999_000_000_000_000));
 
-    // move on to the next cycle after unlock (N+5)
+    await expectAccountToBe(
+      network,
+      Accounts.WALLET_1.stxAddress,
+      100_000_000_000_000 - 80_000_000_000_000 - fee * 3,
+      80_000_000_000_000
+    );
+
+    // Wait until unlock happens that is block N+5 where N is the height of the next reward phase
     await waitForNextRewardPhase(network, orchestrator, 5);
 
     poxInfo = await getPoxInfo(network);
-    // Assert that the current cycle has 920m STX locked and earning
     expect(poxInfo.current_cycle.id).toBe(2);
-    expect(poxInfo.current_cycle.stacked_ustx).toBe(920_000_000_000_000);
-    expect(poxInfo.current_cycle.min_threshold_ustx).toBe(76_670_000_000_000);
+    expect(poxInfo.current_cycle.min_threshold_ustx).toBe(89_920_000_000_000);
+    // Assert that the next cycle has 10790m STX locked
+    // Alice amount was not unlocked because the same
+    // pox address was used as pool address
+    // thereby, the total locked of 1079m was above the minimum of 89m.
+    expect(poxInfo.current_cycle.stacked_ustx).toBe(1_079_000_000_000_000);
 
-    // Alice's table entry has been removed
-    // Check Bob's table entry
+    // Check Alice's table entry
     poxAddrInfo0 = await readRewardCyclePoxAddressListAtIndex(network, 2, 0);
-    expect(poxAddrInfo0?.["total-ustx"]).toEqual(uintCV(920_000_000_000_000));
+    expect(poxAddrInfo0?.["total-ustx"]).toEqual(uintCV(80_000_000_000_000));
 
+    // Check Alice's second table entry
     poxAddrInfo1 = await readRewardCyclePoxAddressListAtIndex(network, 2, 1);
-    expect(poxAddrInfo1?.["total-ustx"]).toBeUndefined();
+    expect(poxAddrInfo1?.["total-ustx"]).toEqual(uintCV(999_000_000_000_000));
 
-    // THIS IS THE BUG
-    // Faucet has no locked STX but Bob's pool still has 920m locked
     await expectAccountToBe(
       network,
-      Accounts.FAUCET.stxAddress,
-      1_000_000_000_000_000 - fee * 2,
-      0
+      Accounts.WALLET_1.stxAddress,
+      100_000_000_000_000 - 80_000_000_000_000 - fee * 3,
+      80_000_000_000_000
     );
   });
 });
